connected engagement summary displays to api results
