+++ src/components/dashboard/widgets/dw_engagement_count.tsx
@@ -1,10 +1,16 @@
 import {
+  AsleepIcon,
+  OnRunningIcon,
+  PendingIcon,
+  TachometerAltIcon,
+} from '@patternfly/react-icons';
+import {
   Button,
   ButtonVariant,
   Card,
   CardBody,
   CardFooter,
-  CardHeader,
+  CardTitle,
   Flex,
   FlexItem,
   Grid,
@@ -13,25 +19,20 @@ import {
   TextContent,
   TextVariants,
 } from '@patternfly/react-core';
-import {
-  AsleepIcon,
-  OnRunningIcon,
-  PendingIcon,
-  TachometerAltIcon,
-} from '@patternfly/react-icons';
 import React, { ComponentClass } from 'react';
-import {
-  Engagement,
-  EngagementStatus,
-  getEngagementStatus,
-} from '../../../schemas/engagement';
+
+// import {
+//     Engagement,
+// } from '../../../schemas/engagement';
 import { SVGIconProps } from '@patternfly/react-icons/dist/js/createIcon';
+import { SummaryCount } from '../../../schemas/engagement';
 
-
-export interface EngagementCountWidgetProps {
-  engagements?: Partial<Engagement>[];
+export interface EngagementCountCardProps {
+  summaryCount?: SummaryCount;
+  isLoading?: boolean;
   onClickCount?: (status: string) => void
 }
+
 interface CountComponentProps {
   count: number;
   icon: ComponentClass<SVGIconProps>;
@@ -67,7 +68,7 @@ const CountComponent = ({ icon: Component, ...props }: CountComponentProps) => {
         >
           <Component size="xl" />
           <TextContent>
-            <Text component={TextVariants.h1}>{props.count}</Text>
+            <Text component={TextVariants.h1}>{props.count || 0}</Text>
             <Text style={{ fontStyle: 'italic', fontSize: '0.8em' }}>
               {props.subtitle}
             </Text>
@@ -78,23 +79,14 @@ const CountComponent = ({ icon: Component, ...props }: CountComponentProps) => {
   )
 };
 
-export const EngagementCountWidget = (
-  props: EngagementCountWidgetProps = {}
-) => {
-  const { engagements = [] } = props;
-  const engagementCounts = [
-    engagements.length,
-    engagements.filter(e => getEngagementStatus(e) === EngagementStatus.active)
-      .length,
-    engagements.filter(
-      e => getEngagementStatus(e) === EngagementStatus.upcoming
-    ).length,
-    engagements.filter(e => getEngagementStatus(e) === EngagementStatus.past || getEngagementStatus(e) === EngagementStatus.terminating
-    ).length,
-  ];
+export function DwEngagementCount({
+  summaryCount,
+  isLoading = false,
+  onClickCount,
+}: EngagementCountCardProps) {
   return (
-    <Card style={{ height: '100%' }}>
-      <CardHeader>
+    <Card isCompact>
+      <CardTitle>
         <TextContent>
           <Text
             component={TextVariants.h2}
@@ -103,9 +95,9 @@ export const EngagementCountWidget = (
             Engagement Summary
           </Text>
         </TextContent>
-      </CardHeader>
+      </CardTitle>
       <CardBody>
-        <Flex
+         <Flex
           alignItems={{ default: 'alignItemsCenter' }}
           style={{ height: '100%', width: '100%' }}
         >
@@ -114,8 +106,8 @@ export const EngagementCountWidget = (
               <Grid hasGutter>
                 <GridItem md={6} lg={3}>
                   <CountComponent
-                    onClickCount={props.onClickCount}
-                    count={engagementCounts[0]}
+                    onClickCount={onClickCount}
+                    count={summaryCount?.any}
                     icon={TachometerAltIcon}
                     status={'all'}
                     subtitle={'All engagements in the system.'}
@@ -124,8 +116,8 @@ export const EngagementCountWidget = (
                 </GridItem>
                 <GridItem md={6} lg={3}>
                   <CountComponent
-                    onClickCount={props.onClickCount}
-                    count={engagementCounts[1]}
+                    onClickCount={onClickCount}
+                    count={summaryCount?.active}
                     icon={OnRunningIcon}
                     status={'active'}
                     subtitle={'Engagements in progress.'}
@@ -134,8 +126,8 @@ export const EngagementCountWidget = (
                 </GridItem>
                 <GridItem md={6} lg={3}>
                   <CountComponent
-                    onClickCount={props.onClickCount}
-                    count={engagementCounts[2]}
+                    onClickCount={onClickCount}
+                    count={summaryCount?.upcoming}
                     icon={PendingIcon}
                     status={'upcoming'}
                     subtitle={'Engagements in the future.'}
@@ -144,8 +136,8 @@ export const EngagementCountWidget = (
                 </GridItem>
                 <GridItem md={6} lg={3}>
                   <CountComponent
-                    onClickCount={props.onClickCount}
-                    count={engagementCounts[3]}
+                    onClickCount={onClickCount}
+                    count={summaryCount?.past + summaryCount?.terminating}
                     icon={AsleepIcon}
                     status={'past'}
                     subtitle={'All completed engagements.'}
@@ -157,7 +149,174 @@ export const EngagementCountWidget = (
           </FlexItem>
         </Flex>
       </CardBody>
-      <CardFooter></CardFooter>
+      <CardFooter />
     </Card>
   );
-};
+}
+
+
+
+
+// import {
+//   Button,
+//   ButtonVariant,
+//   Card,
+//   CardBody,
+//   CardFooter,
+//   CardHeader,
+//   Flex,
+//   FlexItem,
+//   Grid,
+//   GridItem,
+//   Text,
+//   TextContent,
+//   TextVariants,
+// } from '@patternfly/react-core';
+// import {
+//   AsleepIcon,
+//   OnRunningIcon,
+//   PendingIcon,
+//   TachometerAltIcon,
+// } from '@patternfly/react-icons';
+// import React, { ComponentClass } from 'react';
+// import {
+//   Engagement,
+//   EngagementStatus,
+//   getEngagementStatus,
+// } from '../../../schemas/engagement';
+// import { SVGIconProps } from '@patternfly/react-icons/dist/js/createIcon';
+
+
+// export interface EngagementCountWidgetProps {
+//   engagements?: Partial<Engagement>[];
+//   onClickCount?: (status: string) => void
+// }
+// interface CountComponentProps {
+//   count: number;
+//   icon: ComponentClass<SVGIconProps>;
+//   status: string;
+//   subtitle: string;
+//   title: string;
+//   onClickCount: (status: string) => void
+// }
+// const CountComponent = ({ icon: Component, ...props }: CountComponentProps) => {
+//   return (
+//     <Flex
+//       direction={{ default: 'column' }}
+//       alignItems={{ default: 'alignItemsCenter' }}
+//       justifyContent={{ default: 'justifyContentCenter' }}
+//       style={{ height: '100%' }}
+//     >
+//       <FlexItem flex={{ default: 'flex_1' }}>
+//         <Button
+//           onClick={() => props.onClickCount(props.status)}
+//           data-cy={`button_${props.status}`}
+//           variant={ButtonVariant.link}
+//         >
+//           <Text component={TextVariants.h4}>
+//             {props.title}
+//           </Text>
+//         </Button>
+//       </FlexItem>
+//       <FlexItem>
+//         <Flex
+//           direction={{ default: 'column' }}
+//           justifyContent={{ default: 'justifyContentFlexEnd' }}
+//           alignItems={{ default: 'alignItemsCenter' }}
+//         >
+//           <Component size="xl" />
+//           <TextContent>
+//             <Text component={TextVariants.h1}>{props.count}</Text>
+//             <Text style={{ fontStyle: 'italic', fontSize: '0.8em' }}>
+//               {props.subtitle}
+//             </Text>
+//           </TextContent>
+//         </Flex>
+//       </FlexItem>
+//     </Flex>
+//   )
+// };
+
+// export const EngagementCountWidget = (
+//   props: EngagementCountWidgetProps = {}
+// ) => {
+//   const { engagements = [] } = props;
+//   const engagementCounts = [
+//     engagements.length,
+//     engagements.filter(e => getEngagementStatus(e) === EngagementStatus.active)
+//       .length,
+//     engagements.filter(
+//       e => getEngagementStatus(e) === EngagementStatus.upcoming
+//     ).length,
+//     engagements.filter(e => getEngagementStatus(e) === EngagementStatus.past || getEngagementStatus(e) === EngagementStatus.terminating
+//     ).length,
+//   ];
+//   return (
+//     <Card style={{ height: '100%' }}>
+//       <CardHeader>
+//         <TextContent>
+//           <Text
+//             component={TextVariants.h2}
+//             data-testid="engagement-count-card-title"
+//           >
+//             Engagement Summary
+//           </Text>
+//         </TextContent>
+//       </CardHeader>
+//       <CardBody>
+//         <Flex
+//           alignItems={{ default: 'alignItemsCenter' }}
+//           style={{ height: '100%', width: '100%' }}
+//         >
+//           <FlexItem flex={{ default: 'flex_1' }}>
+//             <TextContent style={{ textAlign: 'center' }}>
+//               <Grid hasGutter>
+//                 <GridItem md={6} lg={3}>
+//                   <CountComponent
+//                     onClickCount={props.onClickCount}
+//                     count={engagementCounts[0]}
+//                     icon={TachometerAltIcon}
+//                     status={'all'}
+//                     subtitle={'All engagements in the system.'}
+//                     title="All Engagements"
+//                   />
+//                 </GridItem>
+//                 <GridItem md={6} lg={3}>
+//                   <CountComponent
+//                     onClickCount={props.onClickCount}
+//                     count={engagementCounts[1]}
+//                     icon={OnRunningIcon}
+//                     status={'active'}
+//                     subtitle={'Engagements in progress.'}
+//                     title="Active Engagements"
+//                   />
+//                 </GridItem>
+//                 <GridItem md={6} lg={3}>
+//                   <CountComponent
+//                     onClickCount={props.onClickCount}
+//                     count={engagementCounts[2]}
+//                     icon={PendingIcon}
+//                     status={'upcoming'}
+//                     subtitle={'Engagements in the future.'}
+//                     title="Upcoming Engagements"
+//                   />
+//                 </GridItem>
+//                 <GridItem md={6} lg={3}>
+//                   <CountComponent
+//                     onClickCount={props.onClickCount}
+//                     count={engagementCounts[3]}
+//                     icon={AsleepIcon}
+//                     status={'past'}
+//                     subtitle={'All completed engagements.'}
+//                     title="Past Engagements"
+//                   />
+//                 </GridItem>
+//               </Grid>
+//             </TextContent>
+//           </FlexItem>
+//         </Flex>
+//       </CardBody>
+//       <CardFooter></CardFooter>
+//     </Card>
+//   );
+// };
